refactor: move getCodeDef() to another modjle
